+++ lib/circleci/coverage_reporter/rubycritic/build_result.rb
@@ -2,7 +2,7 @@ require_relative '../abstract_result'
 
 module CircleCI
   module CoverageReporter
-    module SimpleCov
+    module RubyCritic
       class BuildResult < AbstractResult
         # @param path [String]
         # @param build [Build]
@@ -14,14 +14,14 @@ module CircleCI
         # @note Implement {AbstractResult#coverage}
         # @return [Float]
         def coverage
-          last_run_json = find_artifact('.last_run.json') or return Float::NAN
-          JSON.parse(last_run_json.body)['result']['covered_percent'].to_f
+          last_run_json = find_artifact('report.json') or return Float::NAN
+          JSON.parse(last_run_json.body)['score'].to_f
         end
 
         # @note Implement {AbstractResult#url}
         # @return [String]
         def url
-          index_html = find_artifact('index.html') or return '#'
+          index_html = find_artifact('overview.html') or return '#'
           index_html.url
         end